+++ js/jquery.terminal-1.11.1.js
@@ -32,7 +32,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Wed, 27 Dec 2017 14:27:57 +0000
+ * Date: Wed, 27 Dec 2017 16:11:32 +0000
  */
 
 /* TODO:
@@ -2797,7 +2797,7 @@
     }
     $.terminal = {
         version: 'DEV',
-        date: 'Wed, 27 Dec 2017 14:27:57 +0000',
+        date: 'Wed, 27 Dec 2017 16:11:32 +0000',
         // colors from http://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -3723,7 +3723,7 @@
         var temp = $('<div class="terminal temp"><div class="cmd"><span cla' +
                      'ss="prompt">&nbsp;</span></div></div>').appendTo('body');
         if (div) {
-            temp.attr('style', div.attr('style'));
+            temp.attr('style', div.attr('style').replace(/display\s*:\s*none;?/, ''));
         }
         var rect = temp.find('span')[0].getBoundingClientRect();
         var result = {